application/scripts/requirements.txt
backcall==0.1.0
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
cloudpickle==1.2.2
Cython==0.29.13
decorator==4.4.0
idna==2.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
jupyter-client==5.3.3
jupyter-core==4.5.0
numpy==1.17.2
pandas==0.25.1
parso==0.5.1
pdf2image==1.9.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pkg-resources==0.0.0
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
pyocr==0.7.2
PyPDF2==1.26.0
pytesseract==0.3.0
python-dateutil==2.8.0
pytz==2019.3
pyzmq==18.1.0
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
tesseract-ocr==0.0.1
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.6
Wand==0.5.7
wcwidth==0.1.7
